style: resolve clippy warnings
- Remove redundant ai_timeout 'as u64' casts (field is already u64).
- Use sort_by_key(Reverse(..)) for the descending score/token sorts.
- Remove the unused PRAskLine::new_with_ai test constructor.

cargo clippy --all-targets is now clean.

+++ src/ai/openai.rs
@@ -40,7 +40,7 @@ impl OpenAiCompatibleHandler {
             settings.openai.api_base.clone()
         };
         let deployment_id = settings.openai.deployment_id.clone();
-        let timeout_secs = settings.config.ai_timeout as u64;
+        let timeout_secs = settings.config.ai_timeout;
 
         let client = Client::builder()
             .timeout(Duration::from_secs(timeout_secs))
